- Moved MethodTupleTest.java to correct test folder - Added test for Tuple.java
